requirements.txt
alembic==1.9.2
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.2
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.12.1
